Abstract random processes
Currently tests, especially the DNA tests, abstract away randomness by providing a custom sequence replacing `Math.random`. This is unfortunately pretty low level and makes the tests difficult to understand and maintain.

To improve this randomness has been abstracted away into an object that selects one of various choices. The tests can then be rewritten so that they provide an object which selects the things they want in a much easier to understand way.