Implement Environments
Environments are a scaffolding for managing the life cycle of multiple creatures. They manage time, health, population, and the state of all of their contained creatures.

Currently environments are simple: they ensure a minimum number of creatures, spawn food, and provide information to the nearest food item to the various creatures. Every so often the two oldest creatures are recombined under the assumption that the oldest creatures are the fittest.