Long genes are improperly serialized
If a gene is exactly 63 characters long then the length is encoded as `00` which violates the implied assumption that a length never ends in `0`. To correct this the logic for encoding and decoding length is slightly modified so that the value following the last `0` is actually one greater than the value it represents in order to ensure that the value is never represented by a `0`.

Another character could have been used to delimit the length from other parts of the encoding but that would require a larger change to the tests and would increase the size of the encoded DNA even for genes that didn't need the extra size.