DNA tree height is unbounded
When creating a random tree the height is restricted in order to ensure the DNA has a limited size. However, replacing a child during a mutation currently generates a new tree whose maximum height does not depend on its position within the current tree.

This adds a new depth property to each node so that the added tree can depend on where it is placed.