DNA grows to arbitrary lengths
Over the course of many mutations DNA can grow to arbitrary lengths as a result of an unbounded number of gene insertions or duplications. There was a feeble attempt to limit the size of DNA during serialization but it's difficult to ensure that limit at that point in the code.

A better solution is to limit the number of genes in a piece of DNA. This, combined with the maximum depth of any expression tree, places an upper bound on DNA size.