Allow creatures to see and eat each other
This is a very rudimentary implementation of attacking other creatures. The major difficulty is that creatures don't have any method to choose when to attack or to know that they are being attacked in order to properly react to either situation. As a result, this is more a proof of concept than what the final design should be.