Support boolean variables in DNA
After adding support for `isMoving` and `isFast` it became immediately clear that boolean variables are both useful and clunky in the current design. This adds support for booleans in DNA and unfortunately requires a minor modification to the format: the output variable must now specify if it is a real or boolean. The input format for DNA now also needs to specify values for both boolean and real variables since type cannot be inferred.

The creature state handling had to change to support the new DNA input format so that data is read with the correct type. It seems, after writing this change, that the intermediate format returned by `deserializeCreature` is no longer needed and should be removed but this change is large enough as is so that can be done at a later time.

While writing this two other issues were uncovered that were not fixed here for the same reason:

1. If no input or output values are available for the DNA to choose from (eg. `inputBooleans` is the empty string) then the DNA should not try to generate such a value.

2. The swap operator mutation should really be replace tree. Currently it is not possible to swap out `true` or a boolean.