Combine `replaceChild` and `swapOperator`
Previously it was not possible to replace the root of a tree with a large variety of different operations. For example, if the condition on a gene was simple true then that condition could not change because the `swapOperator` operation could not replace true.

This replaces that operation with `replaceTree` which replaces the operation with one whose output is compatible with the original operation and which reuses the existing children, if possible. This mutation subsumes the work of `replaceChild` as well since `replaceChild` essentially performed the same logic but against a child rather than against the root of the tree.