Ensure a minimum number of both creature types
Carnivores naturally have a more difficult time early in the simulation because attacking is more difficult than foraging and because there isn't as much to forage for. This gives herbivores a head start and, if they become successful, they can crowd out spawning space for new carnivores to develop.

Now the simulation ensures a minimum number of each type in order to prevent this issue.