Reproduction is too costly
Both sexual and asexual reproduction are too costly in their current implementations. For carnivores, the system is a loss since the egg is placed on the parent and so will almost always be consumed immediately. Further, reproduction has a lower priority than attacking and so a carnivore is likely to attack a potential mate.

Finally, asexual reproduction tends to quickly kill the creatures that use it because of the significant health loss to both the parent and the child.

To correct these issues parents can no longer eat their own eggs. The cost of asexual reproduction is now the same no matter the parent's health. A parent cannot reproduce if it cannot meet this health cost. The children born of such eggs start at the normal health.